requirements.txt
flask==3.1.2
flask-cors==6.0.1
numpy==2.3.4
pandas==2.3.3
scikit-learn==1.7.2
tensorflow==2.20.0
firebase-admin==6.5.0
python-dotenv==1.0.1